+++ views/motion/_set_proposed_procedure.php
@@ -251,7 +251,7 @@ $votingBlocks = $motion->getMyConsultation()->votingBlocks;
     <h3><?= Yii::t('amend', 'proposal_notify_text') ?></h3>
     <div class="row proposalFrom">
         <?php
-        $replyTo            = \app\components\mail\Tools::getDefaultReplyTo($motion->getMyConsultation());
+        $replyTo            = \app\components\mail\Tools::getDefaultReplyTo($motion->getMyConsultation(), \app\models\db\User::getCurrentUser());
         $fromName           = \app\components\mail\Tools::getDefaultMailFromName($motion->getMyConsultation());
         $placeholderReplyTo = Yii::t('amend', 'proposal_notify_replyto') . ': ' . $replyTo;
         $placeholderName    = Yii::t('amend', 'proposal_notify_name') . ': ' . $fromName;